Jupyter notebook of L2(Error)
This notebook estimates the error in the MFS calculation of the field scattered of a spherical sound hard body. The metric used is the L2 norm of the difference between the scattered field computed according to MFS, and that computed according to an exact harmonic solution of the scattered field. Parameters such as number of source points used, number of quadrature points used (points over which the force integral is computed), and depth of source points, are all varied in order to understand the parameter space within which a scattering calculation may be considered valid.