Force Map Generation
This notebook generates data used to visualize the forces on the x-z plane for two acoustically interacting particles. It plots this force data as a heatmap for various size parameters, and is useful for gauging the relative strength of forces between two particles, as well as the overall spatial structure of the force at various sizes.

This notebook can also be used to save the force map (and location) data to csv files if desired.