Molecular Dynamics Simulation
This notebook performs a molecular dynamics style simulation through the use of a step wise integrator (Runge-Kutte 5/4) that solves the differential equation for a group of acoustically interacting particles. It runs a number of trials, and can be used to save the simulation data for each trial in a specified folder.